Split Project Into Modules, Add Scribe Support
